Notable revision of some fit tactic functionality.
Fit tactics 2 through 4 were not using the proper values for maximum
extensions.  As a result tactic 4 also needed to be revised to function in a
strictly "shrink-to-fit" manner.  (The transformations of tactics 3 and 4 are
properly conceived of as reductions of the original tile extensions.)

Also replacing certain generated images with new ones using the updated code
for truth in advertising.

Maybe it's just me but these corrections seem to result in slightly less
pleasing output than before.  (Possibly because it, if indeed the changes are
functioning as intended, is not resorting as quickly to a minimal tile when
one is specified.  May want to review the fit tactics further.  We may at
least wish to create an alternative (area comparison?) to tactic 2's current
summation rule.)